refactor: stepper constant info extracted to new file
